requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
sqlalchemy==2.0.30
bcrypt==4.1.3
python-dotenv==1.0.1
werkzeug==3.0.1
gunicorn==21.2.0
google-cloud-vision==3.7.2
google-cloud-storage==2.17.0
psycopg2-binary==2.9.6
whitenoise==6.5.0

transformers==4.43.2
torch==2.4.0
reportlab==4.2.0
python-docx==1.1.2
deep-translator==1.11.4
googletrans==4.0.0-rc1
sentencepiece